+++ addons/core/ui/BaseControls.hpp
@@ -1,0 +1,36 @@
+class RscObject;
+class RscText;
+class RscFrame;
+class RscLine;
+class RscProgress;
+class RscPicture;
+class RscPictureKeepAspect;
+class RscVideo;
+class RscHTML;
+class RscButton;
+class RscShortcutButton;
+class RscEdit;
+class RscCombo;
+class RscListBox;
+class RscListNBox;
+class RscXListBox;
+class RscTree;
+class RscSlider;
+class RscXSliderH;
+class RscActiveText;
+class RscActivePicture;
+class RscActivePictureKeepAspect;
+class RscStructuredText;
+class RscToolbox;
+class RscControlsGroup;
+class RscControlsGroupNoScrollbars;
+class RscControlsGroupNoHScrollbars;
+class RscControlsGroupNoVScrollbars;
+class RscButtonTextOnly;
+class RscButtonMenu;
+class RscButtonMenuOK;
+class RscButtonMenuCancel;
+class RscButtonMenuSteam;
+class RscMapControl;
+class RscMapControlEmpty;
+class RscCheckBox;
